track index of entries in key-map
renamed `BinaryHeap::indices` to `key`.
The `Hole` struct is now also responsible for tracking the position
of each entry in the key map. This is done by updating it, every time
`move_to` or `Drop` is invoked.
`BinaryHeap::pop` removes the `key` for the first element from the `key`
map and `BinaryHeap::push` inserts the new key.

At this point, we should have a working binary heap, with push and pop.
That said, all tests are still using the old API and are therefor
failing and we don't have any mutation methods for elements that are
not the first.